Implement Entity Class
Renamed the class from Entites to Entity since the design document had conflicting naming and Entity made more sense in this case. Added a Position class that stores a region, x, and y. This is mainly important so that Entities can hold references to other Entities Positions not just have their own unique location. Added the necessary setters and getters and constructors. Overrode the Position equality operator to support actual comparisons of positions.